python-cli/requirements.txt
typer[all]
rich
questionary
python-dotenv
requests


# Run this to install:
pip install -r requirements.txt